# PERF: tz_localize(None) from dst, asvs