--- a/asv_bench/benchmarks/tslibs/tz_convert.py
+++ b/asv_bench/benchmarks/tslibs/tz_convert.py
@@ -9,7 +9,7 @@ from .tslib import _sizes, _tzs
 class TimeTZConvert:
     params = (
         _sizes,
-        _tzs,
+        [x for x in _tzs if x is not None],
     )
     param_names = ["size", "tz"]